feat: greatly improved hyprland config
Updated the layer animations so that now the animations are more
reasonable. Notifications slide in from the right, fuzzel pops into the
center, and waybar and hyprpaper fade in. I also updated the power
controls so now I have to press `Mq` and hold `M`, then press `[qwert]`
to activate that power option. If I let go of `M` at any point during
that process, then the keychord stops and returns to default binds.